feat(DisacountRepo) : add getValueBytype method to disacount repo for fetch dis value , refactor pricing service by separate applyDisacunt method into two methods ,using disacount value stored in db instead of use hard code vaulues

+++ Domain/Entities/Disacount.cs
@@ -7,7 +7,7 @@ namespace Domain.Entities
     {
          public int Id {  get; set; }
         public string Type { get; set; }
-        public float  Percentag { get; set; }
+        public decimal  Percentag { get; set; }
         public bool IsDeleted { get ; set; }
     }
 }
